split <App> to <App> and <AppRoot>
